server/requirements.docker.txt
numpy
python-dotenv
pydub
aiohttp
requests
openai
groq
datasets
rapidfuzz
edge-tts
